DepthDiff operations. Refacoring.
Added a depthdiff class to common. Now, for the exchanges which don't
natively support the depth diff operation, we will only need to keep
around the old book to generate depth diff messages on the fly. Once
all exchanges support depth diff feeds, the plan will be to subscribe
to all of them in order to create a unified bitcoin book.